requirements.txt
beautifulsoup4==4.12.3
newspaper3k==0.2.8
langdetect==1.0.9
transformers==4.44.2
torch>=2.3
sentencepiece
faiss-cpu==1.8.0.post1
pandas==2.2.2
pyarrow==17.0.0
datasets==2.20.0
huggingface_hub==0.24.6
requests==2.32.3
tqdm==4.66.5
python-dateutil==2.9.0.post0
